+++ src/lang/typechecking.mli
@@ -1,4 +1,5 @@
 exception No_method of string * Type.t
+exception Top_level_override of string * Pos.t option
 
 val debug : bool ref
 val display_types : bool ref
@@ -14,6 +15,7 @@ val type_of_pat :
 
 val check :
   ?env:Typing.env ->
+  check_top_level_override:bool ->
   throw:(bt:Printexc.raw_backtrace -> exn -> unit) ->
   Term.t ->
   unit